farmaproyect-api/requirements.txt
Django==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
django-cors-headers==4.3.1
